File contents was switched
- The class GeoPointMappingDescriptor was in GeoShapeMappingDescriptor
- The class GeoShapeMappingDescriptor was in GeoPointMappingDescriptor
